backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
telethon==1.28.5
python-dotenv==1.0.0
aiohttp==3.8.4
bcrypt==4.0.1
redis==5.0.1
httpx==0.25.1 